+++ src/main/java/vendingmachine/domain/merchandise/MerchandiseCost.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.merchandise;
 
 public class MerchandiseCost {
